backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.4.0
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
python-jose==3.3.0
typing-extensions>=4.12.2
email-validator==2.1.0.post1
Pillow==10.1.0
bcrypt==3.2.2
